Reduction of parameters for drawParagraph method 2nd try
Refactorization for the method drawParagraph in order to reduce the parameters it recieves. There are no tests for the class in which it is implemented, so sonar is going to be checked.

This is the second try in which I'm trying to not create a serious duplication of code, in order to get a successful sonar status